+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 ## 1.11.0 [unreleased]
 
+### Features
+1. [#152](https://github.com/influxdata/influxdb-client-python/pull/152): WriteApi supports generic Iterable type
+
 ### API
- 1. [#151](https://github.com/influxdata/influxdb-client-python/pull/151): Default port changed from 9999 -> 8086
+1. [#151](https://github.com/influxdata/influxdb-client-python/pull/151): Default port changed from 9999 -> 8086
 
 ## 1.10.0 [2020-08-14]