+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 1.11.0 [unreleased]
+## 1.11.0 [2020-10-02]
 
 ### Features
 1. [#152](https://github.com/influxdata/influxdb-client-python/pull/152): WriteApi supports generic Iterable type